Move Nav related component tests to it's own subdirectory
