+++ client/src/layouts/specs/Nav/Nav.spec.tsx
@@ -5,9 +5,9 @@ import { HashRouter as Router, Route, Redirect } from 'react-router-dom';
 import { render, cleanup, fireEvent } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
 
-import configureStore from '../../store/configureStore'
+import configureStore from '../../../store/configureStore'
 
-import Nav from '../Nav/Nav';
+import Nav from '../../Nav/Nav';
 
 
 describe('A component that renders a navbar', () => {
