+++ client/src/layouts/specs/Nav/TodoAppBar.spec.tsx
@@ -5,10 +5,10 @@ import { HashRouter as Router, Route, Redirect } from 'react-router-dom';
 import { render, cleanup, fireEvent } from '@testing-library/react';
 import '@testing-library/jest-dom/extend-expect';
 
-import configureStore from '../../store/configureStore';
+import configureStore from '../../../store/configureStore';
 
 
-import TodoAppBar from '../Nav/TodoAppBar';
+import TodoAppBar from '../../Nav/TodoAppBar';
 
 const store = configureStore();
